Refactor store to include history & local storage
