+++ src/store/index.js
@@ -1,15 +1,5 @@
-import { createStore, applyMiddleware } from "redux"
-import { composeWithDevTools } from "redux-devtools-extension"
-import promiseMiddleware from "redux-promise-middleware"
-import logger from "redux-logger"
+import createStore from "./store"
+import history from "./history"
+import { loadFromLocalStorage, saveToLocalStorage } from "./local-storage"
 
-const middleware = [promiseMiddleware]
-const composeEnhancers = composeWithDevTools({})
-
-if (process.env.NODE_ENV !== "production") {
-  middleware.push(logger)
-}
-
-export default reducers => {
-  return createStore(reducers, composeEnhancers(applyMiddleware(...middleware)))
-}
+export { createStore, history, loadFromLocalStorage, saveToLocalStorage }
